tests: load minitest before rr
When running the tests via rake/hoe (for example, with the `rake test`
command), Rake invokes ruby and loads minitest/autorun before invoking
the any of the actual test suite files.

On the other hand, when running the tests outside of rake/hoe, the test
suite's helper.rb loads the RR gem before loading minitest/autorun. This
ordering causes the test suite to fail because it cannot recognize the
"mock()" function from RR.

Switch the ordering so that minitest/autorun loads before rr. This
allows the test suite to pass in both cases.