Update safelist.rb to include overflow-x and y
Just run into the problem of overflow-x and y disappearing. 
It seems like the end-goal is to not to have to maintain a list like this, but at least in my use case overflow-y is used quite a bit. 
So it would be pretty great to have it as allowed property too!