Extract Loofah::HTML5::Scrub.allowed_uri?
which was previously coupled with Nokogiri in `scrub_uri_attribute` and
so not reusable.

Also, fix a bug where an allowed data URI mediatype would be rejected
if it didn't have a semicolon (i.e., key/value parameters or a base64
extension), for example: 'data:text/plain,hello world'

Also, fix a bug where a string starting with a number could be treated
as a URI schema (e.g., `1foo:`). URI schemes must start with a letter,
not a digit (per RFC 3986).